choose hw to build from cmd line

+++ build.sh
@@ -1,7 +1,8 @@
 #!/bin/bash
+[ -z $1 ] && { echo "Usage: $0 N , where N is the hw number in [1..6]" ; exit 1 ; }
 rm -fr build
 mkdir build
 cd build
-cmake ../
+HW="HW$1" cmake ../
 make
